// src/shared/enums/DocumentType.enum.ts
export enum DocumentTypeEnum {
  TYPE_DRIVER_LICENCE_FRONT = 'TYPE_DRIVER_LICENCE_FRONT',
  TYPE_DRIVER_LICENCE_BACK = 'TYPE_DRIVER_LICENCE_BACK',
  TYPE_VEHICLE_REGISTRATION_STICKER = 'TYPE_VEHICLE_REGISTRATION_STICKER',
  TYPE_INSURANCE_POLICY_PDF = 'TYPE_INSURANCE_POLICY_PDF',
  TYPE_DRIVER_PHOTO = 'TYPE_DRIVER_PHOTO',
  TYPE_VEHICLE_IMAGE = 'TYPE_VEHICLE_IMAGE',
  TYPE_VEHICLE_REGISTRATION_IMAGE = 'TYPE_VEHICLE_REGISTRATION_IMAGE',
  TYPE_VEHICLE_PHOTO = 'TYPE_VEHICLE_PHOTO',
  TYPE_INSURANCE_CERTIFICATE_PICTURE_FILE = 'TYPE_INSURANCE_CERTIFICATE_PICTURE_FILE',
  TYPE_VEHICLE_IDENTIFICATION_NUMBER = 'TYPE_VEHICLE_IDENTIFICATION_NUMBER',
}